1.0 patch 23 fix delete Method and add batch add room

+++ src/main/java/cse/ooad/project/service/TeacherService.java
@@ -6,16 +6,14 @@ import com.opencsv.CSVReader;
 import cse.ooad.project.model.*;
 import cse.ooad.project.repository.*;
 
-import java.io.DataInputStream;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.InputStreamReader;
+import java.io.*;
 import java.nio.charset.StandardCharsets;
 import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Objects;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -72,11 +70,17 @@ public class TeacherService {
 
     @Transactional
     public Boolean deleteFloor(Long id) {
+        if (floorRepository.findById(id).isPresent()) {
+            for (Room room : floorRepository.findById(id).get().getRoomList()) {
+                boolean f=deleteRoom(room.getRoomId());
+                if(!f) return false;
+            }
+        }
         return floorRepository.removeByFloorId(id) != 0;
     }
 
     public Floor updateFloor(Floor floor) {
-         return floorRepository.save(floor);
+        return floorRepository.save(floor);
     }
 
     public Region saveRegion(Region region) {
@@ -86,11 +90,17 @@ public class TeacherService {
 
     @Transactional
     public Boolean deleteRegion(Long id) {
+        if (regionRepository.findById(id).isPresent()) {
+            for (Building building : regionRepository.findById(id).get().getBuildingList()) {
+                boolean f=deleteBuilding(building.getBuildingId());
+                if(!f) return false;
+            }
+        }
         return regionRepository.deleteByRegionId(id) != 0;
     }
 
     public Region updateRegion(Region region) {
-       return regionRepository.save(region);
+        return regionRepository.save(region);
     }
 
     public Room saveRoom(Room room) {
@@ -101,10 +111,26 @@ public class TeacherService {
         return roomRepository.save(room);
     }
 
+    @Transactional
+    public Boolean deleteRoomStarList(Long roomId) {
+        Room room = roomRepository.getRoomsByRoomId(roomId);
+        if (room == null) {
+            return false;
+        }
+        List<Group> groupList = room.getGroupStarList();
+        for (Group group : groupList) {
+            group.getRoomStarList().remove(room);
+            groupRepository.save(group);
+        }
+        return true;
+    }
 
     @Transactional
     public Boolean deleteRoom(Long id) {
-         return roomRepository.deleteByRoomId(id) != 0;
+        if (deleteRoomStarList(id)) {
+            return roomRepository.deleteByRoomId(id) != 0;
+        }
+        return false;
     }
 
     public Building saveBuilding(Building building) {
@@ -118,6 +144,12 @@ public class TeacherService {
 
     @Transactional
     public Boolean deleteBuilding(Long id) {
+        if (buildingRepository.findById(id).isPresent()) {
+            for (Floor floor : buildingRepository.findById(id).get().getFloorList()) {
+                boolean f=deleteFloor(floor.getFloorId());
+                if(!f) return false;
+            }
+        }
         return buildingRepository.removeByBuildingId(id) != 0;
     }
 
@@ -127,25 +159,25 @@ public class TeacherService {
 
 
     //以学生为单位调换宿舍
-    public Boolean transRoom(Long id1, Long id2){
+    public Boolean transRoom(Long id1, Long id2) {
         Student student1 = studentRepository.getStudentByStudentId(id1);
         Student student2 = studentRepository.getStudentByStudentId(id2);
-        if (!Objects.equals(student1.getType(), student2.getType())){
+        if (!Objects.equals(student1.getType(), student2.getType())) {
             return false;
         }
 
         Group group1 = student1.getGroup();
         Group group2 = student2.getGroup();
 
-        if (group1 != null){
+        if (group1 != null) {
             student2.setGroupId(group1.getGroupId());
-            if (Objects.equals(group1.getLeader(), student1.getStudentId())){
+            if (Objects.equals(group1.getLeader(), student1.getStudentId())) {
                 group1.setLeader(student2.getStudentId());
             }
         }
-        if (group2 != null){
+        if (group2 != null) {
             student1.setGroupId(group2.getGroupId());
-            if (Objects.equals(group2.getLeader(), student2.getStudentId())){
+            if (Objects.equals(group2.getLeader(), student2.getStudentId())) {
                 group2.setLeader(student1.getStudentId());
             }
         }
@@ -162,8 +194,8 @@ public class TeacherService {
         try {
             DataInputStream in = new DataInputStream(new FileInputStream(file));
             CSVReader csvReader = new CSVReader(new InputStreamReader(in, StandardCharsets.UTF_8),
-                CSVParser.DEFAULT_SEPARATOR,
-                CSVParser.DEFAULT_QUOTE_CHARACTER, CSVParser.DEFAULT_ESCAPE_CHARACTER, 1);
+                    CSVParser.DEFAULT_SEPARATOR,
+                    CSVParser.DEFAULT_QUOTE_CHARACTER, CSVParser.DEFAULT_ESCAPE_CHARACTER, 1);
             String[] strs;
             List<Student> list = new ArrayList<>();
             List<Password> passwordList = new ArrayList<>();
@@ -188,12 +220,51 @@ public class TeacherService {
         }
     }
 
+
+    public void batchSaveRoomNew(File file) {
+        try {
+            DataInputStream in = new DataInputStream(new FileInputStream(file));
+            CSVReader csvReader = new CSVReader(new InputStreamReader(in, StandardCharsets.UTF_8),
+                    CSVParser.DEFAULT_SEPARATOR,
+                    CSVParser.DEFAULT_QUOTE_CHARACTER, CSVParser.DEFAULT_ESCAPE_CHARACTER, 1);
+            String[] strs;
+
+            List<String> nameList = new ArrayList<>();
+            List<Integer> typeList = new ArrayList<>();
+            List<String> introList = new ArrayList<>();
+            List<String> floorNameList = new ArrayList<>();
+            List<String> buildingNameList = new ArrayList<>();
+            List<String> regionNameList = new ArrayList<>();
+
+            while ((strs = csvReader.readNext()) != null) {
+                nameList.add(strs[6]);
+                introList.add(strs[7]);
+                typeList.add(Integer.parseInt(strs[8]));
+                floorNameList.add(strs[4]);
+                buildingNameList.add(strs[2]);
+                regionNameList.add(strs[0]);
+            }
+            csvReader.close();
+            roomRepository.batchInsertRoomDataFunction(
+                    nameList.toArray(new String[0]),
+                    typeList.toArray(new Integer[0]),
+                    introList.toArray(new String[0]),
+                    floorNameList.toArray(new String[0]),
+                    buildingNameList.toArray(new String[0]),
+                    regionNameList.toArray(new String[0])
+            );
+
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
     public void batchSaveRoom(File file) {
         try {
             DataInputStream in = new DataInputStream(new FileInputStream(file));
             CSVReader csvReader = new CSVReader(new InputStreamReader(in, StandardCharsets.UTF_8),
-                CSVParser.DEFAULT_SEPARATOR,
-                CSVParser.DEFAULT_QUOTE_CHARACTER, CSVParser.DEFAULT_ESCAPE_CHARACTER, 1);
+                    CSVParser.DEFAULT_SEPARATOR,
+                    CSVParser.DEFAULT_QUOTE_CHARACTER, CSVParser.DEFAULT_ESCAPE_CHARACTER, 1);
             String[] strs;
 
             HashMap<String, Region> regionHashMap = new HashMap<>();
@@ -205,10 +276,10 @@ public class TeacherService {
                 Region region = new Region();
                 region.setName(strs[0]);
                 region.setIntro(strs[1]);
-                if(regionHashMap.get(region.getName()) == null){
+                if (regionHashMap.get(region.getName()) == null) {
                     region = regionRepository.save(region);
                     regionHashMap.put(region.getName(), region);
-                }else {
+                } else {
                     region = regionHashMap.get(region.getName());
                 }
 
@@ -217,10 +288,10 @@ public class TeacherService {
                 building.setName(strs[2]);
                 building.setIntro(strs[3]);
                 building.setRegionId(region.getRegionId());
-                if (buildingHashMap.get(region.getName() + building.getName()) == null){
+                if (buildingHashMap.get(region.getName() + building.getName()) == null) {
                     building = buildingRepository.save(building);
                     buildingHashMap.put(region.getName() + building.getName(), building);
-                }else {
+                } else {
                     building = buildingHashMap.get(region.getName() + building.getName());
                 }
 
@@ -230,11 +301,11 @@ public class TeacherService {
                 floor.setIntro(strs[5]);
                 floor.setBuildingId(building.getBuildingId());
 
-                if (floorHashMap.get(region.getName()+building.getName()+floor.getName()) == null){
+                if (floorHashMap.get(region.getName() + building.getName() + floor.getName()) == null) {
                     floor = floorRepository.save(floor);
-                    floorHashMap.put(region.getName()+building.getName()+floor.getName(), floor);
-                }else {
-                    floor = floorHashMap.get(region.getName()+building.getName()+floor.getName());
+                    floorHashMap.put(region.getName() + building.getName() + floor.getName(), floor);
+                } else {
+                    floor = floorHashMap.get(region.getName() + building.getName() + floor.getName());
                 }
 
                 Room room = new Room();
@@ -244,11 +315,10 @@ public class TeacherService {
                 room.setStatus(Integer.parseInt(strs[9]));
                 room.setFloorId(floor.getFloorId());
                 Timestamp timestamp = new Timestamp(System.currentTimeMillis());
-                Comment comment=new Comment(null,strs[6],null,room.getRoomId(),0L,timestamp,false);
-                commentRepository.save(comment);
+                Comment comment = new Comment(null, strs[6], null, room.getRoomId(), 0L, timestamp, false);
+                comment = commentRepository.save(comment);
                 room.setCommentBaseId(comment.getCommentId());
                 rooms.add(room);
-
             }
             roomRepository.saveAll(rooms);
             csvReader.close();
